Move fixture one level higher
